+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/GroovyLanguageServer.java
@@ -20,13 +20,13 @@ import com.google.common.io.Files;
 import com.palantir.ls.groovy.compilation.GroovyTreeParser;
 import com.palantir.ls.groovy.compilation.GroovyWorkspaceCompiler;
 import com.palantir.ls.groovy.compilation.GroovycWrapper;
+import com.palantir.ls.groovy.server.StreamLanguageServerLauncher;
 import com.palantir.ls.groovy.services.GroovyTextDocumentService;
 import com.palantir.ls.groovy.services.GroovyWindowService;
 import com.palantir.ls.groovy.services.GroovyWorkspaceService;
-import com.palantir.ls.server.StreamLanguageServerLauncher;
-import com.palantir.ls.util.GroovyConstants;
-import com.palantir.ls.util.Uris;
-import com.palantir.ls.util.WorkspaceUriSupplier;
+import com.palantir.ls.groovy.util.GroovyConstants;
+import com.palantir.ls.groovy.util.Uris;
+import com.palantir.ls.groovy.util.WorkspaceUriSupplier;
 import io.typefox.lsapi.InitializeParams;
 import io.typefox.lsapi.InitializeResult;
 import io.typefox.lsapi.LanguageDescription;
