+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/LanguageServerState.java
@@ -16,7 +16,7 @@
 
 package com.palantir.ls.groovy;
 
-import com.palantir.ls.api.CompilerWrapper;
+import com.palantir.ls.groovy.api.CompilerWrapper;
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.MessageParams;
 import io.typefox.lsapi.PublishDiagnosticsParams;
