+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/compilation/GroovyTreeParser.java
@@ -22,11 +22,11 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import com.google.common.base.Optional;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import com.palantir.ls.api.TreeParser;
 import com.palantir.ls.groovy.CompilationUnitProvider;
-import com.palantir.ls.util.Ranges;
-import com.palantir.ls.util.UriSupplier;
-import com.palantir.ls.util.Uris;
+import com.palantir.ls.groovy.api.TreeParser;
+import com.palantir.ls.groovy.util.Ranges;
+import com.palantir.ls.groovy.util.UriSupplier;
+import com.palantir.ls.groovy.util.Uris;
 import io.typefox.lsapi.Location;
 import io.typefox.lsapi.ReferenceParams;
 import io.typefox.lsapi.SymbolInformation;
