+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/compilation/GroovyWorkspaceCompiler.java
@@ -23,13 +23,13 @@ import com.google.common.base.Throwables;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.google.common.io.Files;
-import com.palantir.ls.api.WorkspaceCompiler;
 import com.palantir.ls.groovy.CompilationUnitProvider;
-import com.palantir.ls.util.DefaultDiagnosticBuilder;
-import com.palantir.ls.util.GroovyConstants;
-import com.palantir.ls.util.Ranges;
-import com.palantir.ls.util.SourceWriter;
-import com.palantir.ls.util.Uris;
+import com.palantir.ls.groovy.api.WorkspaceCompiler;
+import com.palantir.ls.groovy.util.DefaultDiagnosticBuilder;
+import com.palantir.ls.groovy.util.GroovyConstants;
+import com.palantir.ls.groovy.util.Ranges;
+import com.palantir.ls.groovy.util.SourceWriter;
+import com.palantir.ls.groovy.util.Uris;
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.DiagnosticSeverity;
 import io.typefox.lsapi.FileEvent;
