+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/services/GroovyTextDocumentService.java
@@ -21,8 +21,8 @@ import static com.google.common.base.Preconditions.checkArgument;
 import com.google.common.base.Optional;
 import com.google.common.collect.Lists;
 import com.palantir.ls.groovy.LanguageServerState;
-import com.palantir.ls.util.Ranges;
-import com.palantir.ls.util.Uris;
+import com.palantir.ls.groovy.util.Ranges;
+import com.palantir.ls.groovy.util.Uris;
 import io.typefox.lsapi.CodeActionParams;
 import io.typefox.lsapi.CodeLens;
 import io.typefox.lsapi.CodeLensParams;
