+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/util/DefaultDiagnosticBuilder.java
@@ -14,15 +14,21 @@
  * limitations under the License.
  */
 
-package com.palantir.ls.util;
+package com.palantir.ls.groovy.util;
 
-public final class GroovyConstants {
+import static com.google.common.base.Preconditions.checkNotNull;
 
-    private GroovyConstants() {
-        // constant class
+import io.typefox.lsapi.DiagnosticSeverity;
+import io.typefox.lsapi.builders.DiagnosticBuilder;
+
+public final class DefaultDiagnosticBuilder extends DiagnosticBuilder {
+
+    public DefaultDiagnosticBuilder(String message, DiagnosticSeverity severity) {
+        checkNotNull(message, "message cannot be null");
+        this.message(message);
+        this.severity(severity);
+        this.range(Ranges.UNDEFINED_RANGE);
+        this.source("groovyc");
     }
-
-    public static final String GROOVY_LANGUAGE_NAME = "groovy";
-    public static final String GROOVY_LANGUAGE_EXTENSION = "groovy";
 
 }
