+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/util/Ranges.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.ls.util;
+package com.palantir.ls.groovy.util;
 
 import static com.google.common.base.Preconditions.checkArgument;
