+++ groovy-language-server/src/test/java/com/palantir/ls/groovy/compilation/GroovycWrapperTest.java
@@ -29,9 +29,9 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.palantir.ls.groovy.CompilationUnitProvider;
 import com.palantir.ls.groovy.DefaultCompilationUnitProvider;
-import com.palantir.ls.util.DefaultDiagnosticBuilder;
-import com.palantir.ls.util.Ranges;
-import com.palantir.ls.util.WorkspaceUriSupplier;
+import com.palantir.ls.groovy.util.DefaultDiagnosticBuilder;
+import com.palantir.ls.groovy.util.Ranges;
+import com.palantir.ls.groovy.util.WorkspaceUriSupplier;
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.DiagnosticSeverity;
 import io.typefox.lsapi.FileChangeType;
