+++ groovy-language-server/src/test/java/com/palantir/ls/groovy/services/GroovyTextDocumentServiceTest.java
@@ -24,11 +24,11 @@ import static org.mockito.Mockito.when;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import com.palantir.ls.api.CompilerWrapper;
 import com.palantir.ls.groovy.GroovyLanguageServerState;
 import com.palantir.ls.groovy.LanguageServerState;
-import com.palantir.ls.util.DefaultDiagnosticBuilder;
-import com.palantir.ls.util.Ranges;
+import com.palantir.ls.groovy.api.CompilerWrapper;
+import com.palantir.ls.groovy.util.DefaultDiagnosticBuilder;
+import com.palantir.ls.groovy.util.Ranges;
 import io.typefox.lsapi.CompletionItemKind;
 import io.typefox.lsapi.CompletionList;
 import io.typefox.lsapi.Diagnostic;
