+++ groovy-language-server/src/test/java/com/palantir/ls/groovy/util/DefaultDiagnosticBuilderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.ls.util;
+package com.palantir.ls.groovy.util;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
